move test to test_fact
